+++ src/dda/config/model/user.py
@@ -14,11 +14,12 @@ class UserConfig(Struct, frozen=True):
     name = "U.N. Owen"
     email = "void@some.where"
     ```
-    These values will be used for dda-related functionality, such as telemetry.
-    Both `email` and `name` can be set to `auto`, in which case they will be equal to the values in the [`[tools.git]`][dda.config.model.tools.GitConfig] section.
+
+    These values will be used for `dda`-related functionality like telemetry. Both `email` and `name` can be
+    set to `auto`, in which case they will be equal to the values in the
+    [`[tools.git.author]`][dda.config.model.tools.GitAuthorConfig] section.
     ///
     """
 
-    # Default username and email are equal to the values in [tools.git]
     name: str = "auto"
     email: str = "auto"