+++ src/dda/tools/uv.py
@@ -5,13 +5,14 @@ from __future__ import annotations
 
 import os
 import sys
+from contextlib import contextmanager
 from functools import cached_property
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
-from dda.tools.base import Tool
+from dda.tools.base import ExecutionContext, Tool
 
 if TYPE_CHECKING:
-    from types import TracebackType
+    from collections.abc import Generator
 
     from dda.utils.fs import Path
     from dda.utils.venv import VirtualEnv
@@ -30,13 +31,24 @@ class UV(Tool):
     This also makes modifying the installation of UV itself safe on Windows.
     """
 
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
+    @contextmanager
+    def execution_context(self, command: list[str]) -> Generator[ExecutionContext, None, None]:
+        if self.path is None:
+            yield ExecutionContext(command=["uv", *command], env_vars={})
+            return
 
-        self.__safe_path: str | None = None
+        import shutil
 
-    def format_command(self, command: list[str]) -> list[str]:
-        return [self.__safe_path or self.path or "uv", *command]
+        from dda.utils.fs import Path
+
+        path = Path(self.path)
+        safe_path = path.with_stem(f"{path.stem}-{path.id}")
+        shutil.copy2(self.path, safe_path)
+
+        try:
+            yield ExecutionContext(command=[str(safe_path), *command], env_vars={})
+        finally:
+            safe_path.unlink()
 
     @cached_property
     def path(self) -> str | None:
@@ -55,21 +67,3 @@ class UV(Tool):
             self.wait(["venv", str(path), "--seed", "--python", sys.executable], message="Creating virtual environment")
 
         return VirtualEnv(path)
-
-    def __enter__(self) -> None:
-        if self.path is not None:
-            import shutil
-
-            from dda.utils.fs import Path
-
-            path = Path(self.path)
-            safe_path = path.with_stem(f"{path.stem}-{path.id}")
-            shutil.copy2(self.path, safe_path)
-            self.__safe_path = str(safe_path)
-
-    def __exit__(
-        self, exc_type: type[BaseException] | None, exc_value: BaseException | None, traceback: TracebackType | None
-    ) -> None:
-        if self.__safe_path is not None:
-            os.remove(self.__safe_path)
-            self.__safe_path = None